requirements.txt
aniso8601==7.0.0
Click==7.0
Flask==1.1.1
flask-marshmallow==0.10.1
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.4.0
itsdangerous==1.1.0
Jinja2==2.10.1
libvirt-python==5.3.0
MarkupSafe==1.1.1
marshmallow==3.0.1
marshmallow-sqlalchemy==0.17.2
mysqlclient==1.4.4
pytz==2019.2
six==1.12.0
SQLAlchemy==1.3.8
Werkzeug==0.15.5
